feat: utill 파일로 재사용성을 높임
